chore: group all object and public_key modules
