+++ src/lib.rs
@@ -5,21 +5,19 @@ use tonic_health::pb::health_server::{Health, HealthServer};
 
 use crate::{
     admin::AdminGrpc,
-    cache::Cache,
     config::Config,
     db::connect_and_migrate,
+    domain::OsError,
     mailbox::MailboxGrpc,
     object::ObjectGrpc,
+    public_key::Cache,
     public_key::PublicKeyGrpc,
     replication::ReplicationState,
     server::health::init_health_service,
     storage::{Storage, new_storage},
-    types::OsError,
 };
 
 pub mod admin;
-pub mod authorization;
-pub mod cache;
 pub mod config;
 pub mod consts;
 pub mod datastore;
@@ -34,7 +32,6 @@ pub mod public_key;
 pub mod replication;
 pub mod server;
 pub mod storage;
-pub mod types;
 
 pub mod pb {
     tonic::include_proto!("objectstore");
@@ -59,7 +56,12 @@ impl AppContext {
     /// 3. Build gRPC services
     pub async fn new(config: Arc<Config>) -> Result<Self, OsError> {
         let db_pool = connect_and_migrate(&config.db).await?;
-        let cache = Cache::new(db_pool.clone()).await?;
+
+        let cache = {
+            let initial_keys = datastore::get_all_public_keys(&db_pool).await?;
+            Cache::new(initial_keys).await?
+        };
+
         let storage = new_storage(&config.storage).await?;
 
         let admin_service = AdminGrpc::new(config.clone());
