+++ src/mailbox.rs
@@ -1,12 +1,12 @@
 use crate::datastore;
 use crate::domain::VecUtil;
+use crate::domain::{GrpcResult, OsError};
 use crate::pb::mailbox_service_server::MailboxService;
 use crate::pb::{AckRequest, GetRequest, MailPayload};
 use crate::proto::UuidUtil;
-use crate::types::{GrpcResult, OsError};
 use crate::{
-    cache::{Cache, PublicKeyState},
     config::Config,
+    public_key::{Cache, PublicKeyState},
 };
 
 use fastrace_macro::trace;
