+++ src/main.rs
@@ -1,7 +1,7 @@
 use object_store::AppContext;
 use object_store::config::Config;
+use object_store::domain::Result;
 use object_store::server::configure_and_start_server;
-use object_store::types::Result;
 
 #[tokio::main]
 async fn main() -> Result<()> {
